reorganized app structure
