+++ src/App.scss
@@ -1,8 +1,8 @@
 @use "./partials/setup";
 @use "./partials/variables";
 @use "./partials/mainStyles";
-@use "./partials/input";
-@use "./partials/genreApi";
+@use "./partials/form";
 @use "./partials/loader";
 @use "./partials/playlist";
+@use "./partials/results";
 @use "./partials/savedPlaylists";
