+++ src/partials/_results.scss
@@ -1,0 +1,4 @@
+.playlistButtons{
+    display: flex;
+    justify-content: space-around;
+}
